+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
     "crates/parser",
     "crates/box",
     "crates/pretty_text",
-    "crates/showcase-cli",
     "demo",
     "examples",
 ]
