--- crates/showcase-cli/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "showcase-cli"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-bevy = "0.16"
-bevy-inspector-egui = "0.32.0"
-bevy_pretty_text = { path = "../pretty_text", version = "0.1.0" }
-bevy_capture = { version = "0.3.0", features = ["gif"] }
-clap = { version = "4.5.42", features = ["derive"] }
-indicatif = "0.18.0"
-
-[build-dependencies]
-bevy_reflect = { version = "0.16", features = ["documentation"] }
